HUE-2534 [core] HTTP Thrift transport implementation
Re-using requests in order to get HTTPS and Kerberos for free.
Implement for HBase Thrift Server v1
Implement for HiveServer2

To activate the mode, set in their hue.ini:
transport_mode=http

To activate https:
With Hive:
Use Hue with REQUEST_CA_BUNDLE cf. HUE-2082

With HBase:
Append https:// to the hostnames in the list of servers