HUE-2727 [spark] Only accept numeric input for numeric spark execution settings
With this commit the spark setting inputs for number of cores and executors are forced integers.

This commit adds a new ko binding handler for numeric text input. The binding can be used instead of the extender when desired, for instance when using the ko mapping plugin without custom mapping configuration. The binding makes use of the existing numeric extender and the ko textInput binding.